requirements.txt
anyio==3.6.2
APScheduler==3.10.1
bandit==1.7.5
certifi==2022.12.7
charset-normalizer==3.1.0
flake8==6.0.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.2
pbr==5.11.1
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.0
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rich==13.3.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stevedore==5.0.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
